Hide stderr coming out of 'git pull'
The download progress of 'git pull' was showing, and ruining the lovely
output, so it was silenced. Git doesn't bother to check the stderr is
open before writing to it so /dev/null was used rather than closing it
prematurely.

For the same reason, stdout is piped through 'cat' in order to ensure
that all 'git pull' output is read. This prevents 'head -n1' causing git
to die of signal 13 (SIGPIPE) when it closes the pipe.